Backend/requirements.txt
asgiref==3.4.1
beautifulsoup4==4.10.0
certifi==2021.5.30
cloudinary==1.26.0
Django==3.2.8
django-bootstrap5==2.1.2
djangorestframework==3.12.4
importlib-metadata==2.1.1
Pillow==8.3.2
psycopg2-binary==2.9.1
pytz==2021.3
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.2
typing-extensions==3.10.0.2
urllib3==1.26.7
zipp==3.6.0
